+++ BE/baseball/src/main/java/com/codesquad/baseball/DTO/TeamDTO.java
@@ -1,5 +1,6 @@
 package com.codesquad.baseball.DTO;
 
+import com.codesquad.baseball.DTO.record.PlayerRecordDTO;
 import com.codesquad.baseball.domain.Team;
 
 import java.util.Comparator;
