+++ BE/baseball/src/main/java/com/codesquad/baseball/DTO/score/GameScoreDTO.java
@@ -1,4 +1,4 @@
-package com.codesquad.baseball.DTO;
+package com.codesquad.baseball.DTO.score;
 
 import com.codesquad.baseball.domain.Game;
