+++ BE/baseball/src/main/java/com/codesquad/baseball/DTO/score/TeamScoreDTO.java
@@ -1,4 +1,4 @@
-package com.codesquad.baseball.DTO;
+package com.codesquad.baseball.DTO.score;
 
 import com.codesquad.baseball.domain.Score;
 import com.codesquad.baseball.domain.Team;
