+++ BE/baseball/src/main/java/com/codesquad/baseball/service/TeamService.java
@@ -1,8 +1,8 @@
 package com.codesquad.baseball.service;
 
-import com.codesquad.baseball.DTO.RequestPlayerRecordDTO;
+import com.codesquad.baseball.DTO.record.request.RequestPlayerRecordDTO;
 import com.codesquad.baseball.DTO.TeamDTO;
-import com.codesquad.baseball.DTO.TeamScoreDTO;
+import com.codesquad.baseball.DTO.score.TeamScoreDTO;
 import com.codesquad.baseball.domain.Player;
 import com.codesquad.baseball.domain.Score;
 import com.codesquad.baseball.domain.Team;
